removir os componentes: CatalogProduct e CardProduct agora eles estao implementando diretamente nas telas com seus respectivo componente view e criei o prototipo da tela Cart
