+++ src/Catalog/style.js
@@ -1,10 +1,10 @@
 import styled from "styled-components";
 
 export const CatalogListContainer = styled.FlatList.attrs({
-  contentContainerStyle: {
-    paddingTop: 80,     
-    paddingBottom: 80,  
-  },
+    contentContainerStyle: {
+        paddingTop: 80,
+        paddingBottom: 80,
+    },
 })`
   flex: 1;
 `;
