+++ src/Product/index.js
@@ -1,9 +1,22 @@
-import { useContext, useEffect } from "react";
+import { useContext, useEffect, useState } from "react";
 import { Text } from 'react-native';
 import { BackgroundContext } from "../context/BackgroundProvider";
 import { home } from "../styles/colors";
 import ProductView from "../components/ProductView";
-import CardProduct from "../components/CardProduct";
+
+import {
+  ButtonAddToCart,
+  ButtonBuy,
+  ProductDescription,
+  ProductPrice,
+  ProductTitle,
+  ProductCardView,
+  ProductImage,
+  ButtonTextBuy,
+  ButtonTextCart
+} from "./style";
+
+import { useNavigation } from "../context/NavigationContext";
 
 export default function Product() {
   const { background, setBackground } = useContext(BackgroundContext);
@@ -12,6 +25,9 @@ export default function Product() {
     setBackground(home.bars_background);
   }, []);
 
+  const { props } = useNavigation();
+  const [selectedProduct] = useState(props.product);
+  console.log(selectedProduct)
   return (
     <ProductView
       footer={true}
@@ -30,7 +46,23 @@ export default function Product() {
       >
         Enviar para CIDADE - CEP
       </Text>
-      <CardProduct />
+      <ProductCardView>
+        <ProductImage source={selectedProduct.image} />
+        <ProductTitle>{selectedProduct.name}</ProductTitle>
+        <ProductPrice>R$ {selectedProduct.price.toFixed(2).replace('.', ',')}</ProductPrice>
+        <ButtonAddToCart>
+          <ButtonTextCart>Adicionar ao Carrinho</ButtonTextCart>
+        </ButtonAddToCart>
+        <ButtonBuy>
+          <ButtonTextBuy>Comprar Agora</ButtonTextBuy>
+        </ButtonBuy>
+        <ProductDescription style={{ fontWeight: "bold" }}>
+          Descrição do produto:
+        </ProductDescription>
+        <ProductDescription>
+          {selectedProduct.description || "Sem descrição disponível"}
+        </ProductDescription>
+      </ProductCardView>
     </ProductView>
   );
 }
